backend/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
Flask-Limiter==3.5.1
Flask-SQLAlchemy==3.1.1
Flask-SocketIO==5.3.6
Flask-JWT-Extended==4.6.0
python-dotenv==1.0.1
dynaconf==3.2.4
itsdangerous==2.2.0
passlib[bcrypt]==1.7.4
cryptography==43.0.1
gunicorn==23.0.0
requests==2.32.3
rich==13.9.2
pytest==8.2.2
pytest-cov==5.0.0
bandit==1.7.9
black==24.8.0
ruff==0.6.9
python-socketio==5.11.0


